Carousel: apply the largeFile fix in place, drop the extracted helper module
The previous commit fixed the portrait black-screen bug by extracting
selectBestImageUrl() and its helpers into a new select-best-image-url.js
module, purely so the URL-selection logic could be unit tested. That
extraction cost more than it was worth for a one-line patch:

- jetpack-carousel.js is served raw (unbundled) under SCRIPT_DEBUG, so the
  new top-level `import` threw "Cannot use import statement outside a
  module" and broke the whole carousel in debug/local environments.
- Working around that forced the enqueue to always load the webpack bundle,
  diverging from trunk (and dropping the jetpack_get_file_for_environment
  filter unless carefully preserved).
- webpack turns every module .js into its own entry, so the split also
  emitted a dead, unenqueued select-best-image-url.min.js bundle that ships
  in the plugin.

Keep the actual fix inline: guard against a falsy (empty-string) largeFile
rather than only `undefined`, so a gallery image with no data-large-file
attribute falls back to the original file instead of returning an empty src.

Remove the extracted module and its unit tests. The functions live inside a
private IIFE and can't be exercised without re-introducing that module, so
the regression test goes with it; the behaviour is instead documented with
an inline comment on the guard.
